+++ src/classification.cpp
@@ -19,7 +19,7 @@ std::ostream& operator<<(std::ostream& out, const cpphots::Features& feats) {
     int w = std::to_string(*maxval).size();
 
     out << "[" << std::setw(w) << std::setfill(' ') << feats[0];
-    for (unsigned int i = 1; i < feats.size(); i++) {
+    for (size_t i = 1; i < feats.size(); i++) {
         out << " " << std::setw(w) << std::setfill(' ') << feats[i];
     }
     out << "]";
@@ -31,7 +31,7 @@ std::ostream& operator<<(std::ostream& out, const cpphots::Features& feats) {
 
 namespace cpphots {
 
-Classifier::Classifier(unsigned int n_classes) {
+Classifier::Classifier(size_t n_classes) {
 
     class_feats.resize(n_classes);
 
@@ -42,7 +42,7 @@ Classifier::Classifier(const std::vector<std::string>& classes)
 
     class_feats.resize(classes.size());
 
-    for (unsigned int i = 0; i < classes.size(); i++) {
+    for (size_t i = 0; i < classes.size(); i++) {
         reverse_class_names.insert({classes[i], i});
     }
 
@@ -51,20 +51,20 @@ Classifier::Classifier(const std::vector<std::string>& classes)
 Classifier::Classifier(Classifier* other)
     :class_feats(other->class_feats), class_names(other->class_names), reverse_class_names(other->reverse_class_names) {}
 
-void Classifier::setClassFeatures(unsigned int cid, const Features& feats) {
+void Classifier::setClassFeatures(size_t cid, const Features& feats) {
     class_feats[cid] = feats;
 }
 
 void Classifier::setClassFeatures(const std::string& cname, const Features& feats) {
-    unsigned int cid = reverse_class_names.at(cname);
+    size_t cid = reverse_class_names.at(cname);
     setClassFeatures(cid, feats);
 }
 
-unsigned int Classifier::classifyID(const Features& feats) const {
+size_t Classifier::classifyID(const Features& feats) const {
 
-    unsigned int argmin = class_feats.size();
+    size_t argmin = class_feats.size();
     double mindist = std::numeric_limits<double>::max();
-    for (unsigned int i = 0; i < class_feats.size(); i++) {
+    for (size_t i = 0; i < class_feats.size(); i++) {
         double d = computeDistance(class_feats[i], feats);
         if (d < mindist) {
             mindist = d;
@@ -81,7 +81,7 @@ std::string Classifier::classifyName(const Features& feats) const {
     if (class_names.empty())
         throw std::runtime_error("Cannot output class name if no names were set at construction time");
 
-    unsigned int cid = classifyID(feats);
+    size_t cid = classifyID(feats);
 
     return class_names[cid];
 
@@ -95,7 +95,7 @@ double StandardClassifier::computeDistance(const Features& f1, const Features& f
     }
 
     double dist = 0.0;
-    for (unsigned int i = 0; i < f1.size(); i++) {
+    for (size_t i = 0; i < f1.size(); i++) {
         dist += std::pow(double(f1[i]) - double(f2[i]), 2);
     }
 
@@ -114,7 +114,7 @@ double NormalizedClassifier::computeDistance(const Features& f1, const Features&
     double card2 = std::accumulate(f2.begin(), f2.end(), 0.0);
 
     double dist = 0.0;
-    for (unsigned int i = 0; i < f1.size(); i++) {
+    for (size_t i = 0; i < f1.size(); i++) {
         dist += std::pow(f1[i] / card1 - f2[i] / card2, 2);
     }
 
@@ -133,7 +133,7 @@ double BhattacharyyaClassifier::computeDistance(const Features& f1, const Featur
     double card2 = std::accumulate(f2.begin(), f2.end(), 0.0);
 
     double dist = 0.0;
-    for (unsigned int i = 0; i < f1.size(); i++) {
+    for (size_t i = 0; i < f1.size(); i++) {
         dist += std::sqrt(f1[i] / card1 * f2[i] / card2);
     }
